+++ src/Network/Network/ChangeLog.md
@@ -24,6 +24,7 @@
 * Onboarded `Microsoft.EventGrid/namespaces` to private link cmdlets
 * Fixed bug in `NewAzureApplicationGatewayFirewallCustomRuleGroupByVariable` to add "GeoLocation" as a valid input for VariableName
 * Added breaking change message for parameter `VariableName` in `NewAzureApplicationGatewayFirewallCustomRuleGroupByVariable` to remove "Geo" as a valid input.
+* Added min scale units and max scale units to support autoscale for Virtual Network gateways
 
 ## Version 6.1.1
 * Onboarded `Microsoft.ElasticSan/elasticSans` to private link cmdlets